hoc2021: set hoc_mode on test to 'soon-hoc'
Other environments (including staging & production) have now moved to `hoc_mode` being `"soon-hoc"` for 2021.  The test environment was left out, but that was helful because it helped make sure that the changes to enable `"soon-hoc"` didn't regress any of the pages while still running as `"pre-hoc"`.  Now, though, we can move the test environment over.