Explicitly define some auto-loaded submodules
Specifically, explicitly define some of the submodules we implicitly use for code organization. According to [this StackOverflow answer](https://stackoverflow.com/questions/64548442/unable-to-autoload-constant-bug-in-rails-5-2-0/64601195#64601195), this is a new requirement in Rails 5.2, and may explain the "Rails unable to autoload constant; expected (some file) to define it" error we've occasionally been seeing when deploying during times of high traffic.