Pegasus: fix for partials with parameters, for translations
When we reviewed pages with translated content, we discovered a problem with the use of `{{ }}`-style partials with parameters in some cases.  The use of `.undump` to unescape the string was throwing an exception when the string contained non-ASCII characters, though that exception was being silenced immediately.

The quick solution here is to no longer use `.undump` to unescape the string, and instead to explicitly remove the surrounding `"`s, and to replace all `\"`s with `"`s.