Reinforce Custom JSON Parsing Logic
We've got some custom JSON-parsing logic in our AWS Secrets Manager interface. Specifically, custom logic which in addition to wrapping parsed JSON Objects in fancy Ruby accessor helpers will nicely handle values which cannot be parsed to JSON by simply returning them directly.

Unfortunately, the current version of the JSON gem we're using does not support parsing scalar values:

```
[development] dashboard > JSON.parse("123")
Traceback (most recent call last):
        1: from (irb):1
JSON::ParserError (784: unexpected token at '123')
[development] dashboard > JSON.parse("1.0")
Traceback (most recent call last):
        1: from (irb):2
JSON::ParserError (784: unexpected token at '1.0')
```

I'm not sure why this is; this behavior is (as far as I can find) undocumented and entirely undesirable. The version of the JSON gem to which we'd like to upgrade as part of the Ruby 2.7 upgrade has the expected behavior:

```
[development] dashboard > JSON.parse("123")
=> 123
[development] dashboard > JSON.parse("1.0")
=> 1.0
[development] dashboard >
```

Which for the most part is a good thing, but does break the behavior here. I'm not entirely sure of the best way to preserve the existing behavior; or, indeed, if we even _want_ to preserve the existing behavior. Open to suggestions!