Fix rare infinite loop
It was possible to get in an infinite loop of
setInstructionsMaxHeightNeeded leading to a InstructionsCSF's
componentDidUpdate which calls setInstructionsMaxHeightNeeded again.

It appeared to involve a very narrow set of instructions in which it's
too narrow to render any text and they get short again.  We then
oscillate between the collapsd and non-collapsed height.