Re-apply `better_errors`, `rambling-trie`, and `honeybadger` Gem Updates
My first pass involved just updating in Gemfile.lock the specific pinned versions, since the Gemfile entries for these gems were nonspecific. Unfortunately, that led to issues during the deploy to production. We're still not sure exactly why, but Chef installed the old versions of the gems (as specified by just the Gemfile and not the Gemfile.lock) and then failed when it couldn't find the new versions of the gems (as specified by the Gemfile.lock).

This time, I updated the Gemfile to explicitly indicate which versions
of these gems we care about for Ruby 2.7 compatibility, then ran

    bundle update better_errors honeybadger rambling-trie